Send form data and other information with API, sprint 11

+++ src/utils/api.js
@@ -61,19 +61,35 @@ class Api {
     })
   }
 
-  likeCard(cardId) {
-    return this._request(`${this._url}/cards/${cardId}/likes`, {
-      method: 'PUT',
-      headers: this._headers,
-    })
-  }
+  changeLikeCardStatus(cardId, isLiked) {
+    if (isLiked) {
+      return this._request(`${this._url}/cards/${cardId}/likes`, {
+        method: 'DELETE',
+        headers: this._headers,
+      })
+    } else {
+      return this._request(`${this._url}/cards/${cardId}/likes`, {
+        method: 'PUT',
+        headers: this._headers,
+      })
+    }
+}
 
-  dislikeCard(cardId) {
-    return this._request(`${this._url}/cards/${cardId}/likes`, {
-      method: 'DELETE',
-      headers: this._headers,
-    })
-  }
+  // // likeCard(cardId) {
+  //   return this._request(`${this._url}/cards/${cardId}/likes`, {
+  //     method: 'PUT',
+  //     headers: this._headers,
+  //   })
+  // }
+
+  // dislikeCard(cardId) {
+  //   return this._request(`${this._url}/cards/${cardId}/likes`, {
+  //     method: 'DELETE',
+  //     headers: this._headers,
+  //   })
+  // }
+
+
 }
 
 const api = new Api ({
